Allow for transient entity-specific state via the effector mechanism
Effector is instantiated in F context upon actor recovery by a synchronous call to the dispatcher. This makes it possible to maintain an in-memory stateful effector, e.g., to implement some caching. Also expose the entity algebra to the effector constructor, so that it can be referenced by these transient stateful structures. This commit also adapts the passivation state handling mechanisms to support this model.