Add support for differentiated run modes for side-effects
A new `runModeFor` method is introduced in `SideEffect`, with the default set to `Async` to ensure backward compatibility. With `Async` mode, it triggers in *run & forget* mode so that command reply is not delayed by any lengthy side-effect. With `Sync` mode, the side-effect runs to completion before the next command is processed by the entity: this can be useful to simplify the side-effect logic as it precludes accounting for concurrency, but can hurt the system's responsiveness.